update resource CRUD routes

+++ src/key/key.service.ts
@@ -2,7 +2,7 @@ import { PaginationOptions } from '@/common/dto/pagination-options.dto';
 import { CreateKeyDto } from '@/key/dto/create-key.dto';
 import { UpdateKeyDto } from '@/key/dto/update-key.dto';
 import { PrismaService } from '@/prisma/prisma.service';
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 
 @Injectable()
 export class KeyService {
@@ -11,7 +11,19 @@ export class KeyService {
   /**
    * Create key
    */
-  create(createKeyDto: CreateKeyDto) {
+  async create(user_id: string, createKeyDto: CreateKeyDto) {
+    const keychain = await this.prisma.keychain.findUnique({
+      where: {
+        id: createKeyDto.keychain_id,
+        user_id,
+      },
+    });
+
+    if (!keychain)
+      throw new NotFoundException(
+        `Keychain ${createKeyDto.keychain_id} not found`,
+      );
+
     return this.prisma.key.create({
       data: createKeyDto,
     });
@@ -20,10 +32,36 @@ export class KeyService {
   /**
    * Find all keys
    */
-  findAll(options?: PaginationOptions) {
+  findAll(user_id: string, paginationOptions?: PaginationOptions) {
     return this.prisma.key.findMany({
-      skip: options.skip,
-      take: options.take,
+      where: {
+        keychain: {
+          user_id,
+        },
+      },
+      skip: paginationOptions.skip,
+      take: paginationOptions.take,
+      orderBy: { name: 'asc' },
+    });
+  }
+
+  /**
+   * Find all keychain keys
+   */
+  findAllByKeychain(
+    user_id: string,
+    keychain_id: string,
+    paginationOptions?: PaginationOptions,
+  ) {
+    return this.prisma.key.findMany({
+      where: {
+        keychain_id,
+        keychain: {
+          user_id,
+        },
+      },
+      skip: paginationOptions.skip,
+      take: paginationOptions.take,
       orderBy: { name: 'asc' },
     });
   }
@@ -31,16 +69,18 @@ export class KeyService {
   /**
    * Find key
    */
-  findOne(id: string) {
-    return this.prisma.key.findUniqueOrThrow({ where: { id } });
+  findOne(id: string, user_id: string) {
+    return this.prisma.key.findUniqueOrThrow({
+      where: { id, keychain: { user_id } },
+    });
   }
 
   /**
    * Update key
    */
-  update(id: string, updateKeyDto: UpdateKeyDto) {
+  update(id: string, user_id: string, updateKeyDto: UpdateKeyDto) {
     return this.prisma.key.update({
-      where: { id },
+      where: { id, keychain: { user_id } },
       data: updateKeyDto,
     });
   }
@@ -48,7 +88,7 @@ export class KeyService {
   /**
    * Remove key
    */
-  remove(id: string) {
-    return this.prisma.key.delete({ where: { id } });
+  remove(id: string, user_id: string) {
+    return this.prisma.key.delete({ where: { id, keychain: { user_id } } });
   }
 }
